+++ frontend/src/components/button/CustomButton.tsx
@@ -1,4 +1,4 @@
-const CustomButton = ({ label, method, cssModifier = 'default' }: {label?: string, method: () => any, cssModifier?: string}) => {
+const CustomButton = ({ label, method, cssModifier = 'default', type = 'button' }: {label?: string, method: () => any, cssModifier?: string, type: "button" | "submit" | "reset" | undefined}) => {
   const inheritedFunction = () => {
     if(!method) return;
     return method()
@@ -8,6 +8,7 @@ const CustomButton = ({ label, method, cssModifier = 'default' }: {label?: strin
       <button
         className="base-button__button"
         onClick={() => inheritedFunction()}
+        type={type}
       >
         { label ?? 'Button' }
       </button>