+++ frontend/src/components/navigation/sideButtons.ts
@@ -14,21 +14,21 @@ const sideBarButtons = [
     icon: 'bi-file-person-fill',
     path: 'borrowers',
   },
-  {
-    name: 'recovery',
-    icon: 'bi-basket3-fill',
-    path: 'recovery',
-  },
-  {
-    name: 'calendar',
-    icon: 'bi-calendar3-week-fill',
-    path: 'calendar',
-  },
-  {
-    name: 'coms',
-    icon: 'bi-chat-right-quote-fill',
-    path: 'communications',
-  },
+  // {
+  //   name: 'recovery',
+  //   icon: 'bi-basket3-fill',
+  //   path: 'recovery',
+  // },
+  // {
+  //   name: 'calendar',
+  //   icon: 'bi-calendar3-week-fill',
+  //   path: 'calendar',
+  // },
+  // {
+  //   name: 'coms',
+  //   icon: 'bi-chat-right-quote-fill',
+  //   path: 'communications',
+  // },
 ];
 
 export default sideBarButtons;